test: rename commonTests.ts to not be included in coverage report
